requirements.txt
apex==0.9.10dev
filelock==3.3.1
gdown==4.2.0
h5py==3.3.0
joblib==1.1.0
kaldiio==2.17.2
librosa==0.8.1
matplotlib==3.3.4
numpy==1.21.2
PyYAML==6.0
scikit_learn==1.0.1
scipy==1.7.1
SoundFile==0.10.3.post1
tensorboardX==2.4.1
tensorflow==2.7.0
torch==1.10.0
tqdm==4.62.3
